complete changeMessage, deleteMessage functionality
