+++ src/view/components/ChatForm/index.tsx
@@ -13,24 +13,26 @@ import { Message } from '../../../bus/messages/types';
 export type FormSubmitEvent = React.FormEvent<HTMLFormElement>;
 export type FormChangeEvent = React.ChangeEvent<HTMLInputElement>;
 
-type PropTypes = {
-    onSubmitForm: (event: FormSubmitEvent, message?: Message) => void;
-    onChangeInput: (event: FormChangeEvent) => void;
+export type FormPropTypes = {
+    onSubmitForm: (event: FormSubmitEvent, message?: Message, username?: string) => void;
+    onChangeInput: (event: FormChangeEvent, message?: Message) => void;
     inputValue: string;
+    message?: Message
 }
 
-export const SubmitForm: FC<PropTypes> = ({ onSubmitForm, onChangeInput, inputValue }) => {
+export const SubmitForm: FC<FormPropTypes> = ({ onSubmitForm, onChangeInput, inputValue, message }) => {
     return (
         <S.Container>
             <S.Form onSubmit = { onSubmitForm }>
                 <S.Input
                     type = 'text'
                     value = { inputValue }
-                    onChange = { onChangeInput }
+                    onChange = { (event) => onChangeInput(event, message) }
                 />
                 <S.SubmitBtn
                     disabled = { !inputValue }
-                    type = 'submit'>send
+                    type = 'submit'>
+                    send
                 </S.SubmitBtn>
             </S.Form>
         </S.Container>
