+++ src/view/components/RegistrationUserInfo/styles.ts
@@ -1,0 +1,27 @@
+// Core
+import styled from 'styled-components';
+
+export const Container = styled.section`
+    flex-direction: column;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+`;
+
+export const DecorIcon = styled.img`
+`;
+
+export const Title = styled.h1`
+    font-family: 'Inter-Black';
+    font-weight: 900;
+    margin-bottom: 50px;
+    text-align: center;
+    font-size: 64px;
+    line-height: 67px;
+`;
+
+export const TitleAccentWord = styled.span`
+    display: block;
+    color: #55BBF8;
+`;
