+++ src/view/components/index.ts
@@ -1,3 +1,5 @@
 export * from './ErrorBoundary';
 export * from './UserInfo';
 export * from './ChatForm';
+export * from './RegistrationUserInfo';
+export * from './RegistrationForm';
